Refactor BootstrapTable into separate modules
* Modules: initialization, body, check, data, detail, header, pagination, search, and toolbar
